app/requirements.txt
Flask==2.0.3
Werkzeug==2.2.2
requests==2.27.1
python-dotenv==0.20.0
openai==1.27.0
